move main.go to a folder
